Commenting and file structure.
